HOTT-2128 Namespaced admin news items controller
